Commit 13
Added in a textbox to display the amount of friends gotten on the
Game.cs page. Added a Globalclass.cs page so that the amount of
'friends' gotten in the Game.cs is saved to the main page. When you
click a button containing this global variable on the Main page, that
global variable is written to the textbox underneath it.